+++ src/FreeElephants/JsonApi/DTO/Exception/JsonApiDtoExceptionInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace FreeElephants\JsonApi\DTO\Exception;
+
+interface JsonApiDtoExceptionInterface
+{
+
+}