+++ src/FreeElephants/JsonApi/DTO/Exception/UnexpectedValueException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace FreeElephants\JsonApi\DTO\Exception;
+
+class UnexpectedValueException extends \UnexpectedValueException implements JsonApiDtoExceptionInterface
+{
+
+}